Added Upgrade Menu PolyGone.java: - updated enemy speed calc - updated enemy spawn rate - added enemy dropped xp variable - added upgrade menu object to setup - moved order of guis and huds - fixed click through bug - fixed instant selction bug
UpgradeMenu.java:
- rewrote framework
- added modularity for upgrade card creation
- moved logic to act method
- updated blurring to exclude debug hud

PauseMenu.java:
- converted button dimensions variables to global
- moved logic to act method

GUI.java:
- updated playerxp counter to double type variable for more accurate tracking

DebugHUD.java:
- updated player xp debug text to double type rounded to 2 decimals for more accurate tracking
- removed unused variables

Player.java:
- improved code conventions by assigning new variables to existing constants
- added console logging of player level updates
- updated bullet speed variable to be public access

Other:
- fixed various bugs
- added comments
- improved code conventions
- rebalanced game

+++ src/main/java/Project/GUI.java
@@ -92,13 +92,13 @@ public class GUI extends GameObject {
         }
     }
 
-    public void drawXPBar(Graphics2D g2d, int x, int y, int currentPlayerXP, int playerXPBarMaxXP) {
+    public void drawXPBar(Graphics2D g2d, int x, int y, double currentPlayerXP, int playerXPBarMaxXP) {
         //xp bar dimensions
         int barWidth = this.getWidth() - (2 * x);
         int barHeight = 35; //thickness of bar
 
         //calculates the width of the xp bar
-        double xpPercentage = (double) currentPlayerXP / playerXPBarMaxXP; //casts to double to avoid rounding to 0
+        double xpPercentage = currentPlayerXP / (double)playerXPBarMaxXP; //casts to double to avoid rounding to 0
         xpPercentage = Math.max(0.0, Math.min(1.0, xpPercentage)); //makes sure the value is between 0 and 1
         int fillWidth = (int) (barWidth * xpPercentage);
